Add files via upload
The Panel Data we used in the panel regression. Since the data size will become quite large by saving as a ".csv" , here we uploaded it in the ".RData" format which provided a effective compressing method. Another reason is that we processed our data mainly by R. The order of the variables (the columns) in the Panel Data is as follows:
1. The Number of Effective Idea Submission (the dependent variable)
2. The Constraint
3. The Neighbours’ Average Effective Idea Submission
4. The Frequency of Interactions
5.  Days Being Observed
6. Number of Neighbours
7. The Number of All Idea Submission